+++ Cargo.toml
@@ -13,23 +13,18 @@ readme = "README.md"
 [dev-dependencies]
 criterion = { version = "0.5.1", features = ["html_reports"] }
 rand = "0.8.5"
+scc = { version = "2.2.5" }
 
 [dependencies]
 serde = { version = "1.0.151", optional = true, features = ["derive"] }
 ftree = { version = "1.1.0", features = ["serde"] }
-arc-swap = { version = "1.7.1" }
-parking_lot = { version = "0.12.3", features = ["send_guard", "arc_lock"] }
-crossbeam-utils = { version = "0.8.20" }
-crossbeam-skiplist = { version = "0.1.3" }
-scc = { version = "2.2.5" }
-bplustree = { version = "0.1.0" }
+parking_lot = { version = "0.12.3", features = ["send_guard", "arc_lock"], optional = true }
+crossbeam-utils = { version = "0.8.20", optional = true }
+crossbeam-skiplist = { version = "0.1.3", optional = true}
 
 [features]
 serde = ["dep:serde"]
-
-[[bin]]
-name = "main"
-path = "src/bin.rs"
+concurrent = ["dep:crossbeam-utils", "dep:crossbeam-skiplist", "dep:parking_lot"]
 
 [[bench]]
 name = "stdlib"
@@ -38,3 +33,4 @@ harness = false
 [[bench]]
 name = "concurrent"
 harness = false
+features = ["concurrent"]
